+++ include/utils.hpp
@@ -8,9 +8,9 @@
 
 struct location
 {
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 };
 
 struct axis
@@ -24,9 +24,9 @@ struct axis
 struct shape
 {
     // Stores the homogeneous coordinates of each vertex.
-    float vectors[MAX_VERTICES_PER_SHAPE][NUMBER_OF_HOMOGENEOUS_COORDS];
+    double vectors[MAX_VERTICES_PER_SHAPE][NUMBER_OF_HOMOGENEOUS_COORDS];
     // The connection matrix will store relationships between vertices
-    float connectivity[MAX_VERTICES_PER_SHAPE][MAX_VERTICES_PER_SHAPE];
+    double connectivity[MAX_VERTICES_PER_SHAPE][MAX_VERTICES_PER_SHAPE];
     // Number of currently occupied vertex spots for the shape.
     int numVertices;
 };
@@ -35,10 +35,10 @@ int addConnectionToShape(struct shape * target, int source, int destination);
 
 int addVertexToShape(struct shape * target, struct location point);
 
-void copyMatrix(float source[][4], float dest[][4], int n);
+void copyMatrix(double source[][4], double dest[][4], int n);
 
-void matMatMult(float operation[4][4], float target[][4], float destination[][4], int n);
+void matMatMult(double operation[4][4], double target[][4], double destination[][4], int n);
 
-void printMat(float mat[][4], int n);
+void printMat(double mat[][4], int n);
 
 void initShape(struct shape * target);
